ml-service/requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
pandas
pillow
seaborn
flask
flask-cors
